requirements.txt
# minimal requirements file
Django==4.1.5
django-environ==0.9.0
django-extensions==3.2.1
#psycopg2==2.9.5
psycopg2-binary
jsonpickle
django-settings-export
django-guardian

# Crispy Forms stuff
django-crispy-forms
crispy-bootstrap5

# Requirements for UW-Django-SAML2
UW-Django-SAML2

# Datatables View

# Web serving stuff
whitenoise

# install jupyterlab if you want to test django code in a notebook environment
# jupyterlab==3.5.2

# testing & devel stuff
coverage
django-debug-toolbar
black

# REST stuff
djangorestframework
markdown
django-filter
djangorestframework-datatables
django-datatables-view

# Bioinformatics stuff
gffutils

# File stuff
pandas
openpyxl
